Relocate app.config to root
