+++ src/context/reducers/modelSelectionReducer.tsx
@@ -1,5 +1,5 @@
 import { ActionMap, UpdatingModelSelectionTypes } from "../actions/types";
-import { PluginModels } from "../../../src/api/app.config"
+import { PluginModels } from "../../app.config"
 
 export type ModelSelection = {
    xrayModel: string,
