+++ src/services/pacs_integration.tsx
@@ -1,5 +1,5 @@
 import { DcmImage, PFDCMResponse, PACSMainResponse, PACSSeries} from "../context/reducers/dicomImagesReducer";
-import { BASE_PACS_FILE_PATH } from '../api/app.config';
+import { BASE_PACS_FILE_PATH } from '../app.config';
 import axios from 'axios';
 
 declare var process: {
